fix(website): :bug: 修复 og 图片无法显示的问题
- 合并 `layout.ts` 和 `page.ts` 中的 Metadata
- 修复在 X 平台，og 图片无法访问的问题
- 为 og 图片增加 alt 描述
