+++ docs/app/[lang]/layout.tsx
@@ -12,84 +12,91 @@ import { Locale } from "@app/_dictionaries/i18n-config";
 import Logo from "@/public/logo.svg";
 import lingoI18nConfig from "../../../i18n.json";
 import { allLanguages } from "../_dictionaries/i18n-config";
-import getConfig from 'next/config';
+import getConfig from "next/config";
 import { I18NConfig } from "next/dist/server/config-shared";
 import LocalSwitch from "../components/local-switch";
 import { GithubIcon } from "lucide-react";
-import { Analytics } from "@vercel/analytics/next"
+import { Analytics } from "@vercel/analytics/next";
+import { Metadata } from "next";
 
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ lang: Locale }>;
+}): Promise<Metadata> {
+  const { lang } = await params;
+  const dictionary = await getDictionary(lang);
 
-    
-export const metadata = {
-  metadataBase: new URL("https://www.variables-xporter.com"),
-  title: {
-    template: "%s - Variables Xporter",
-    default: "Variables Xporter - Export Figma Variables to shadcn/ui & Tailwind CSS",
-  },
-  description: "Export Figma Variables as production-ready shadcn/ui theme variables, Tailwind CSS configuration files, or CSS Variables. Support for multi-mode design systems.",
-  applicationName: "Variables Xporter",
-  generator: "Next.js",
-  keywords: [
-    "Figma plugin",
-    "design tokens",
-    "shadcn/ui",
-    "Tailwind CSS",
-    "CSS variables",
-    "design system",
-    "variables export",
-    "design to code",
-  ],
-  authors: [{ name: "Kinsey" }],
-  creator: "Kinsey",
-  publisher: "Variables Xporter",
-  appleWebApp: {
-    title: "Variables Xporter",
-    statusBarStyle: "default",
-    capable: true,
-  },
-  formatDetection: {
-    telephone: false,
-  },
-  icons: {
-    icon: "/favicon.ico",
-    shortcut: "/favicon-16x16.png",
-    apple: "/apple-touch-icon.png",
-  },
-  manifest: "/site.webmanifest",
-  other: {
-    "msapplication-TileImage": "/mstile-150x150.png",
-    "msapplication-TileColor": "#fff",
-    "msapplication-config": "/browserconfig.xml",
-  },
-  openGraph: {
-    type: "website",
-    siteName: "Variables Xporter",
-    url: "https://www.variables-xporter.com",
-    title: "Variables Xporter - Export Figma Variables to shadcn/ui & Tailwind CSS",
-    description: "Export Figma Variables as production-ready shadcn/ui theme variables, Tailwind CSS configuration files, or CSS Variables. Support for multi-mode design systems.",
-  },
-  twitter: {
-    card: "summary_large_image",
-    site: "@kinsey57534670",
-    creator: "@kinsey57534670",
-    title: "Variables Xporter - Export Figma Variables to shadcn/ui & Tailwind CSS",
-    description: "Export Figma Variables as production-ready shadcn/ui theme variables, Tailwind CSS configuration files, or CSS Variables.",
-  },
-  robots: {
-    index: true,
-    follow: true,
-    googleBot: {
+  const title = "Variables Xporter";
+  const description = dictionary.pluginHighlight;
+  const url = "https://www.variables-xporter.com";
+
+  return {
+    title: {
+      template: "%s - Variables Xporter",
+      default: "Variables Xporter",
+    },
+    description,
+    keywords: [
+      "Figma plugin",
+      "design tokens",
+      "shadcn/ui",
+      "Tailwind CSS",
+      "CSS variables",
+      "design system",
+      "variables export",
+      "design to code",
+    ],
+    authors: [{ name: "Kinsey" }],
+    creator: "Kinsey",
+    publisher: "Variables Xporter",
+    appleWebApp: {
+      title: "Variables Xporter",
+      statusBarStyle: "default",
+      capable: true,
+    },
+    formatDetection: {
+      telephone: false,
+    },
+    icons: {
+      icon: "/favicon.ico",
+      shortcut: "/favicon-16x16.png",
+      apple: "/apple-touch-icon.png",
+    },
+    manifest: "/site.webmanifest",
+    other: {
+      "msapplication-TileImage": "/mstile-150x150.png",
+      "msapplication-TileColor": "#fff",
+      "msapplication-config": "/browserconfig.xml",
+    },
+    openGraph: {
+      type: "website",
+      siteName: title,
+      url,
+      description,
+    },
+    twitter: {
+      card: "summary_large_image",
+      site: "@kinsey57534670",
+      creator: "@kinsey57534670",
+      description,
+    },
+    robots: {
       index: true,
       follow: true,
-      "max-video-preview": -1,
-      "max-image-preview": "large",
-      "max-snippet": -1,
+      googleBot: {
+        index: true,
+        follow: true,
+        "max-video-preview": -1,
+        "max-image-preview": "large",
+        "max-snippet": -1,
+      },
     },
-  },
-  verification: {
-    google: "google-verification-code",
-  },
-};
+    verification: {
+      google: "google-verification-code",
+    },
+  };
+}
 
 export default async function RootLayout({ children, params }) {
   const { lang } = await params;
@@ -107,7 +114,7 @@ export default async function RootLayout({ children, params }) {
     name: allDictionaries[locale].language,
   }));
 
-  console.log(i18n)
+  console.log(i18n);
 
   const navbar = (
     <Navbar
@@ -137,24 +144,24 @@ export default async function RootLayout({ children, params }) {
   return (
     <html lang={lang as Locale} suppressHydrationWarning>
       <Head
-        // color={{
-        //   hue: {
-        //     light: 242,
-        //     dark: 242,
-        //   },
-        //   saturation: {
-        //     light: 50,
-        //     dark: 100,
-        //   },
-        //   lightness: {
-        //     light: 46,
-        //     dark: 83,
-        //   },
-        // }}
-        // backgroundColor={{
-        //   light: "#fcfcfc",
-        //   dark: "#111111",
-        // }}
+      // color={{
+      //   hue: {
+      //     light: 242,
+      //     dark: 242,
+      //   },
+      //   saturation: {
+      //     light: 50,
+      //     dark: 100,
+      //   },
+      //   lightness: {
+      //     light: 46,
+      //     dark: 83,
+      //   },
+      // }}
+      // backgroundColor={{
+      //   light: "#fcfcfc",
+      //   dark: "#111111",
+      // }}
       />
       <body>
         <Layout
