+++ docs/app/[lang]/page.tsx
@@ -20,75 +20,6 @@ import { getDictionary } from "../_dictionaries/get-dictionary";
 import { Locale } from "../_dictionaries/i18n-config";
 import Link from "next/link";
 import { MagicCard } from "@/components/magicui/magic-card";
-import { Metadata } from "next";
-
-export async function generateMetadata({
-  params,
-}: {
-  params: Promise<{ lang: Locale }>;
-}): Promise<Metadata> {
-  const { lang } = await params;
-  const dictionary = await getDictionary(lang);
-  
-  const title = "Variables Xporter";
-  const description = dictionary.pluginHighlight;
-  const url = "https://www.variables-xporter.com";
-
-  return {
-    title,
-    description,
-    keywords: [
-      "Figma plugin",
-      "design tokens",
-      "shadcn/ui",
-      "Tailwind CSS",
-      "CSS variables",
-      "design system",
-      "variables export",
-      "design to code",
-    ],
-    authors: [{ name: "Kinsey" }],
-    creator: "Kinsey",
-    publisher: "Variables Xporter",
-    robots: {
-      index: true,
-      follow: true,
-      googleBot: {
-        index: true,
-        follow: true,
-        "max-video-preview": -1,
-        "max-image-preview": "large",
-        "max-snippet": -1,
-      },
-    },
-    openGraph: {
-      type: "website",
-      locale: lang,
-      url,
-      title,
-      description,
-      siteName: "Variables Xporter",
-    },
-    twitter: {
-      card: "summary_large_image",
-      title,
-      description,
-      creator: "@kinsey57534670",
-      site: "@kinsey57534670",
-    },
-    alternates: {
-      canonical: url,
-      languages: {
-        en: `${url}/en`,
-        zh: `${url}/zh`,
-        fr: `${url}/fr`,
-      },
-    },
-    verification: {
-      google: "google-verification-code",
-    },
-  };
-}
 
 export default async function IndexPage({
   params,
